requirements.txt
Django==2.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
